+++ com.splunk.modularinput.java.behavior/src/com/splunk/modularinput/java/UnterminatedConditionalException.java
@@ -1,0 +1,10 @@
+package com.splunk.modularinput.java;
+
+public class UnterminatedConditionalException extends Exception {
+	public String tokenName;
+	
+	public UnterminatedConditionalException(String missingTokenName) {
+		super("Unterminated conditional on token \"" + missingTokenName + "\"");
+		tokenName = missingTokenName;
+	}
+}