+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Splunk plug-in for Eclipse
 
+## Version 0.8
+
+* Added support for creating modular inputs in Java.
+
+## Version 0.2.1
+
+* Fixed errors in dependency versions so plug-in works on versions of Eclipse back to 3.6.2.
+* Upgraded the Splunk SDK for Java version 1.2.1.
+	
 ## Version 0.2
 
 * Upgraded to Splunk SDK for Java version 1.2.